Get rid of types.ts files, add background/utils/signing, attempt to follow the desired dependency flow of redux => background services
