+++ background/redux-slices/signing.ts
@@ -1,20 +1,46 @@
 import { createSelector, createSlice } from "@reduxjs/toolkit"
 import Emittery from "emittery"
-import { SiweMessage } from "siwe"
-import { SignTypedDataRequest } from "../../services/signing/types"
 import {
-  EIP4361Data,
-  Events,
   ExpectedSigningData,
   SignDataMessageType,
   SignDataRequest,
-  SigningState,
-  SignOperation,
-} from "./types"
-import { createBackgroundAsyncThunk } from "../utils"
-import { EnrichedSignTypedDataRequest } from "../../services/enrichment"
+  SigningMethod,
+  SignTypedDataRequest,
+} from "../utils/signing"
+import { createBackgroundAsyncThunk } from "./utils"
+import { EnrichedSignTypedDataRequest } from "../services/enrichment"
+import { EIP712TypedData, HexString } from "../types"
+
+type SignOperation<T> = {
+  request: T
+  signingMethod: SigningMethod
+}
+
+type Events = {
+  requestSignTypedData: {
+    typedData: EIP712TypedData
+    account: HexString
+    signingMethod: SigningMethod
+  }
+  requestSignData: {
+    signingData: ExpectedSigningData
+    messageType: SignDataMessageType
+    rawSigningData: string
+    account: HexString
+    signingMethod: SigningMethod
+  }
+  signatureRejected: never
+}
 
 export const signingSliceEmitter = new Emittery<Events>()
+
+type SigningState = {
+  signedTypedData: string | undefined
+  typedDataRequest: EnrichedSignTypedDataRequest | undefined
+
+  signedData: string | undefined
+  signDataRequest: SignDataRequest | undefined
+}
 
 export const initialState: SigningState = {
   typedDataRequest: undefined,
@@ -56,56 +82,6 @@ export const signData = createBackgroundAsyncThunk(
     })
   }
 )
-
-const checkEIP4361: (message: string) => EIP4361Data | undefined = (
-  message
-) => {
-  try {
-    const siweMessage = new SiweMessage(message)
-    return {
-      domain: siweMessage.domain,
-      address: siweMessage.address,
-      statement: siweMessage.statement,
-      version: siweMessage.version,
-      chainId: siweMessage.chainId,
-      expiration: siweMessage.expirationTime,
-      nonce: siweMessage.nonce,
-    }
-  } catch (err) {
-    // console.error(err)
-  }
-
-  return undefined
-}
-
-/**
- * Takes a string and parses the string into a ExpectedSigningData Type
- *
- * EIP4361 standard can be found https://eips.ethereum.org/EIPS/eip-4361
- */
-export const parseSigningData: (signingData: string) => {
-  data: ExpectedSigningData
-  type: SignDataMessageType
-} = (signingData) => {
-  const data = checkEIP4361(signingData)
-  if (data) {
-    return {
-      data,
-      type: SignDataMessageType.EIP4361,
-    }
-  }
-
-  // data = checkOtherType(lines)
-  // if (!!data) {
-  // return data
-  // }
-
-  // add additional checks for any other types to add in the future
-  return {
-    data: signingData,
-    type: SignDataMessageType.EIP191,
-  }
-}
 
 const signingSlice = createSlice({
   name: "signing",
