+++ ui/pages/PersonalSignDetailPanel.tsx
@@ -1,7 +1,7 @@
 import {
   EIP4361Data,
   SignDataMessageType,
-} from "@tallyho/tally-background/redux-slices/signing/types"
+} from "@tallyho/tally-background/utils/signing"
 import { selectSigningData } from "@tallyho/tally-background/redux-slices/signing"
 import React from "react"
 import { EIP191Info, EIP4361Info } from "../components/SignData"
